week_8/day_5/ex-xp/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-phonenumber-field==6.0.0
Faker==9.8.3
phonenumbers==8.12.37
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
